# optimization pass for keybasefriends loading