--- a/go/client/cmd_team.go
+++ b/go/client/cmd_team.go
@@ -30,6 +30,7 @@ func NewCmdTeam(cl *libcmdline.CommandLine, g *libkb.GlobalContext) cli.Command 
 			newCmdTeamDelete(cl, g),
 			newCmdTeamAPI(cl, g),
 			newCmdTeamSettings(cl, g),
+			newCmdTeamProfileLoad(cl, g),
 		},
 	}
 }