… go/teams/chain_test.go
… go/teams/service_helper.go
… go/teams/sig.go
… protocol/avdl/keybase1/teams.avdl
… protocol/js/rpc-gen.js
… protocol/json/keybase1/teams.json
… shared/constants/types/rpc-gen.js