Based on your branch https://github.com/keybase/client/pull/12494. This isn't any faster, but maybe a bit of cleanup.

Rough times for `keybase team profile-load keybasefriends` with warm caches:
```
6s - this PR
9s - this PR with moveState changed to DeepCopy
6s - base PR
22s - master
```

Btw it's now spending more time in verifyLink than in applyNewLink. (Look for TimeBucket). Which did [not used to be true](https://github.com/keybase/client/pull/12550), so that's good.
```
2018-06-25T17:05:24.351996-04:00 ▶ [DEBU keybase util.go:531] 1181a + TeamLoader.load2ILR [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:24.352022-04:00 ▶ [DEBU keybase timetracer.go:52] 1181b | TeamLoader.load2ILR:init [time=25.158µs] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:24.352049-04:00 ▶ [DEBU keybase timetracer.go:52] 1181c | TeamLoader.load2ILR:cache load [time=1.481µs] [tags:LT=oJwuuzpCL5uo,TM=Of2rMV4dNhUd]
2018-06-25T17:05:24.404211-04:00 ▶ [DEBU keybase timetracer.go:52] 11838 | TeamLoader.load2ILR:merkle [time=52.137767ms] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:24.404237-04:00 ▶ [DEBU keybase timetracer.go:52] 11839 | TeamLoader.load2ILR:backfill [time=1.32µs] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:24.404292-04:00 ▶ [DEBU keybase timetracer.go:52] 1183b | TeamLoader.load2ILR:pre-fetch [time=31.589µs] [tags:LT=oJwuuzpCL5uo,TM=Of2rMV4dNhUd]
2018-06-25T17:05:25.499508-04:00 ▶ [DEBU keybase timetracer.go:52] 11841 | TeamLoader.load2ILR:fetch [time=1.095190357s] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:25.807325-04:00 ▶ [DEBU keybase timetracer.go:52] 11842 | TeamLoader.load2ILR:unpack [time=307.777187ms] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:30.494932-04:00 ▶ [DEBU keybase timebuckets.go:48] 1fdb8 TimeBucket CachedUPAKLoader.LoadKeyV2 [time=1.386510149s] [tags:LT=oJwuuzpCL5uo,TM=Of2rMV4dNhUd]
2018-06-25T17:05:30.494946-04:00 ▶ [DEBU keybase timebuckets.go:48] 1fdb9 TimeBucket TeamLoader.verifyLink [time=3.834675663s] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:30.494956-04:00 ▶ [DEBU keybase timebuckets.go:48] 1fdba TimeBucket TeamLoader.applyNewLink [time=735.674506ms] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:30.494967-04:00 ▶ [DEBU keybase timetracer.go:52] 1fdbb | TeamLoader.load2ILR:linkloop (3393) [time=4.687599991s] [tags:LT=oJwuuzpCL5uo,TM=Of2rMV4dNhUd]
2018-06-25T17:05:30.494981-04:00 ▶ [DEBU keybase timetracer.go:52] 1fdbc | TeamLoader.load2ILR:pco [time=1.063µs] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:31.989723-04:00 ▶ [DEBU keybase timetracer.go:52] 20290 | TeamLoader.load2ILR:checkproofs [time=1.494729868s] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:32.154099-04:00 ▶ [DEBU keybase timetracer.go:52] 20295 | TeamLoader.load2ILR:secrets [time=164.348012ms] [tags:LT=oJwuuzpCL5uo,TM=Of2rMV4dNhUd]
2018-06-25T17:05:32.158845-04:00 ▶ [DEBU keybase timetracer.go:52] 20296 | TeamLoader.load2ILR:namecalc [time=4.692809ms] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:32.177796-04:00 ▶ [DEBU keybase timetracer.go:52] 20297 | TeamLoader.load2ILR:put [time=18.906233ms] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:32.177831-04:00 ▶ [DEBU keybase timetracer.go:52] 20298 | TeamLoader.load2ILR:notify [time=1.038µs] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:32.177853-04:00 ▶ [DEBU keybase timetracer.go:62] 20299 | TeamLoader.load2ILR:postcheck [time=2.358µs] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
2018-06-25T17:05:32.177877-04:00 ▶ [DEBU keybase loader.go:684] 2029a - TeamLoader.load2ILR [time=7.825880289s] [tags:TM=Of2rMV4dNhUd,LT=oJwuuzpCL5uo]
```

Please expand `chain.go` which GH collapses.

Instead of CheapCopying the state, the transition functions 'consume' the state object and just re-use its memory for their return values. The transitions funcs are `AppendChainLink` and `InflateLink`. This PR's changes from your branch are basically a long winded version of this:
```
new := old.CheapCopy()
-- becomes --
new = old
old = nil
```

Tried to comment and make sure that once a `TeamSigchainState` is passed into a transition function there are no live references into the object. Because it was deep copied solely for this purpose earlier. This part you already did.

This PR has more code change than it really needs to be because I wanted to get rid of this persistent threadsafe `TeamSigChainPlayer` idea. In favor of just having transition funcs. Internally they use a `teamSigchainPlayer` but nobody needs to know that, it's basically a namespace for helpers.